third_party/WebKit/LayoutTests/virtual/webrtc-wpt-unified-plan/external/wpt/webrtc/RTCRtpTransceiver-setDirection-expected.txt

This is a testharness.js-based test.
FAIL setDirection should change transceiver.direction transceiver.setDirection is not a function
FAIL setDirection with same direction should have no effect transceiver.setDirection is not a function
FAIL setDirection should change transceiver.direction independent of transceiver.currentDirection assert_equals: expected "recvonly" but got "inactive"
Harness: the test ran to completion.
